Implementation of the urlRequest pkg, update of main.go
The urlRequest pkg can be used in order to build a correct url
from an object of mapping.
The main pkg contains two main functions:
-makeRequest, which takes a string to a url and modifies a
pointer in which it will store the answer to the request
-handleRequest, which will build the url, call the makeReqest
function, parse to value to a float and pass this float to a
channel, allowing the function to run in concurrency mode.